Remove snprintf absense workaround hack.
I tried to build access on a Slackware 3.2 again. The result was not very good.

Since then, it was required more code to *modify* (not I say it is not possible
again, but things changed, and access went to use more modern API slightly).
For example, strftime is plain stupid on libc4. And there is no "%zu" printf specifier.
And there is no "%s" strftime format for Unix seconds.
So, it *may* work, but including code which 99% of times will be not used here produces bloat.
So I am removing the last compatibility shims to support *such* old systems.

If your system lacks (v)snprintf, getgrouplist, `date +%s`, "%zu" and SIZE_MAX,
then it is probably a time to upgrade, yep?