Remove "%set nocommands"; -C now only disables closing file descriptors.
"%set nocommands" was silly. It broke su compat program included and it was known.
Hiding "public" information is stupid anyway even if it is buried somewhere else.

-C now only disables closing file descriptors and available only to superuser.
It does not take arguments. It was not widely used by ordinary users during my experience.

Fix delay variable type to one which is accepted by usleep(3): useconds_t.